refactor(streams): move helm streaming core behind EventSink
