+++ apps/desktop/src-tauri/Cargo.toml
@@ -28,6 +28,7 @@ tauri-plugin-dialog = "2"
 srelens-capability = { path = "../../../crates/capability" }
 srelens-mcp = { path = "../../../crates/mcp" }
 srelens-kube = { path = "../../../crates/kube" }
+srelens-streams = { path = "../../../crates/streams" }
 tokio = { version = "1", features = ["rt-multi-thread", "io-std", "io-util", "macros"] }
 # Blocking HTTP for Toolbox tool downloads (run inside spawn_blocking).
 # rustls-no-provider matches tauri-plugin-updater's reqwest so no second TLS
