+++ crates/streams/src/lib.rs
@@ -2,6 +2,7 @@
 //! server: each manager drives srelens-kube streams and emits events into an
 //! [`EventSink`] implemented by the host (Tauri events, WebSocket frames).
 
+pub mod helm;
 pub mod sink;
 pub mod test_util;
